abstracting out read simulation
